requirements.txt
requests==2.30.0
parsel==1.8.1
django==4.2.1
celery==5.2.7
django-bulk-update-or-create==0.3.0
django-celery-beat==2.5.0
python-decouple==3.8
psycopg2-binary==2.9.6
redis==4.5.5
django-redis==5.2.0
retry==0.9.2
gunicorn==20.1.0
google-auth==2.17.3
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0